Better tracking by Makefile for incremental builds
This allows us to disable prep by SBY and manage preprocessing manually
without enormous time cost.

Ideally we would still have the spec SV be pre-compiled, as this is still
slow. It also looks as if there may be some combinational loops going on
too, which might be more difficult to deal with.