Construct an aiger slowly, then tweak it quickly
Yosys (via build_all.ys) will constuct an aiger file for all properties,
which we then modify (aig-manip) to convert assertions to assumptions,
or delete them altogether. Such manipulation is effectively instant in
Rust.

Proof is then managed by insert-property.py, which searches for the best
configuration of rIC3 for that purpose.