[dv/formal] Lean proof of Observational Correctness
Introduces a new proof of the observational correctness property in the
theorem proving language Lean 4. The final proof is in Main.lean. This
required minimal changes to the verification code, though some properties
were added for convenience.

In essence, the top level SV property are written directly as axioms in
Lean, in which we can infer higher level properties.

See the ObservationalCorrectness theorem for the final result: that
the specification and Ibex will produce the same sequence of memory
results forever.

All axioms are explicit, though there are a couple of interest (see the
source).

This also introduces a new CI job, to check that the Lean code is still
correct.